requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
DateTime==5.5
discord==2.3.2
discord-webhook==1.4.1
discord.py==2.5.2
frozenlist==1.5.0
idna==3.10
load-dotenv==0.1.0
multidict==6.2.0
propcache==0.3.1
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
schedule==1.2.2
setuptools==78.1.0
sgfmill==1.1.1
thread==2.0.5
typing_extensions==4.13.0
urllib3==2.3.0
yarl==1.18.3
zope.interface==7.2
